authentication: add Helmholtz AAI service
Motivation:

The Helmholtz AAI service is a Identity and Authorisation
Management (IAM) system which arbitrates authenticated
access to registered services in the context of the Helmholtz
Assosiation. Integrating Helmholtz AAI service into the
Helmholtz Marketplace is in alignment with the goal of this
project.

Modification:

- authenticate -> able to get token
- logout
- enable csrf support

Result:

User can now be authenticate using Unity IDM.